Move interfaces that have no ties to the engine to mod code
