+++ OpenRA.Mods.RA/Traits/Infiltration/InfiltrateForSupportPower.cs
@@ -9,6 +9,7 @@
  */
 #endregion
 
+using OpenRA.Mods.Common.Traits;
 using OpenRA.Primitives;
 using OpenRA.Traits;
